int to unsigned long...
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@3744 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ examples/nbody/nbody/server/unigrid_mesh.cpp
@@ -590,7 +590,7 @@ namespace hpx { namespace components { namespace nbody { namespace server
           {
             
 
-                for (int p=0; p < par->bilist[i].size() ; ++p)
+                for (unsigned long p=0; p < par->bilist[i].size() ; ++p)
                 {
                     
                     j = par->bilist[i][p];
